refactor in progress
