+++ app/Command.hs
@@ -1,19 +1,18 @@
 module Command
   ( module Command,
     module Commands.Extra,
-    module Commands.Html,
-    module Commands.Compile,
-    module Commands.Dev,
+    module Commands.Html.Options,
+    module Commands.Compile.Options,
+    module Commands.Dev.Options,
     module Commands.Doctor,
   )
 where
 
-import Commands.Compile
-import Commands.Dev
+import Commands.Compile.Options
+import Commands.Dev.Options
 import Commands.Doctor
 import Commands.Extra
-import Commands.Html
-import GlobalOptions
+import Commands.Html.Options
 import Juvix.Prelude hiding (Doc)
 import Options.Applicative
 
@@ -21,19 +20,38 @@ data Command
   = Check
   | Compile CompileOptions
   | Html HtmlOptions
-  | Dev InternalCommand
+  | Dev DevCommand
 
-data CommandGlobalOptions = CommandGlobalOptions
-  { _cliCommand :: Command,
-    _cliGlobalOptions :: GlobalOptions
-  }
+commandCheck :: Mod CommandFields Command
+commandCheck =
+  command "typecheck" $
+    info
+      (pure Check)
+      (progDesc "Type check a Juvix file")
 
-makeLenses ''CommandGlobalOptions
+commandCompile :: Mod CommandFields Command
+commandCompile =
+  command "compile" $
+    info
+      (Compile <$> parseCompile)
+      (progDesc "Compile a Juvix file")
 
-parseCommandGlobalOptions :: Parser CommandGlobalOptions
-parseCommandGlobalOptions = do
-  opts <- parseGlobalFlags False
-  cmd <-
+commandHtml :: Mod CommandFields Command
+commandHtml =
+  command "html" $
+    info
+      (Html <$> parseHtml)
+      (progDesc "Generate HTML for a Juvix file")
+
+commandDev :: Mod CommandFields Command
+commandDev =
+  command "dev" $
+    info
+      (Dev <$> parseDevCommand)
+      (progDesc "Commands for the developers")
+
+parseCompilerCommand :: Parser Command
+parseCompilerCommand =
     hsubparser
       ( mconcat
           [ commandGroup "Compiler commands:",
@@ -44,54 +62,3 @@ parseCommandGlobalOptions = do
             commandDev
           ]
       )
-  return (cmd {_cliGlobalOptions = opts <> cmd ^. cliGlobalOptions})
-
-commandCheck :: Mod CommandFields CommandGlobalOptions
-commandCheck =
-  command "typecheck" $
-    info
-      (addGlobalOptions (pure Check))
-      (progDesc "Type check a Juvix file")
-
-commandCompile :: Mod CommandFields CommandGlobalOptions
-commandCompile =
-  command "compile" $
-    info
-      (addGlobalOptions (Compile <$> parseCompile))
-      (progDesc "Compile a Juvix file")
-
-commandHtml :: Mod CommandFields CommandGlobalOptions
-commandHtml =
-  command "html" $
-    info
-      (addGlobalOptions (Html <$> parseHtml))
-      (progDesc "Generate HTML for a Juvix file")
-
-commandDev :: Mod CommandFields CommandGlobalOptions
-commandDev =
-  command "internal" $
-    info
-      (addGlobalOptions (Dev <$> parseInternalCommand))
-      (progDesc "Internal subcommands")
-
---------------------------------------------------------------------------------
--- Misc
---------------------------------------------------------------------------------
-
-cmdDefaultOptions :: Command -> CommandGlobalOptions
-cmdDefaultOptions _cliCommand =
-  CommandGlobalOptions {_cliGlobalOptions = mempty, ..}
-
-liftParserCmd :: Parser Command -> Parser CommandGlobalOptions
-liftParserCmd cmd = cmdDefaultOptions <$> cmd
-
-addGlobalOptions :: Parser Command -> Parser CommandGlobalOptions
-addGlobalOptions parser = do
-  flags1 <- parseGlobalFlags True
-  ~(opts2, _cliCommand) <- addParser (parseGlobalFlags True) parser
-  fs <- parserInputFiles
-  return
-    CommandGlobalOptions
-      { _cliGlobalOptions = flags1 <> opts2 <> mempty {_globalInputFiles = fs},
-        ..
-      }
