+++ app/Commands/Base.hs
@@ -1,0 +1,12 @@
+module Commands.Base
+  ( module App,
+    module GlobalOptions,
+    module Juvix.Compiler.Pipeline,
+    module Juvix.Prelude,
+  )
+where
+
+import App
+import GlobalOptions
+import Juvix.Compiler.Pipeline
+import Juvix.Prelude
