+++ app/Commands/Dev/Termination/CallGraph/Options.hs
@@ -1,0 +1,24 @@
+module Commands.Dev.Termination.CallGraph.Options where
+
+import Data.Text qualified as Text
+import Juvix.Prelude
+import Options.Applicative
+
+newtype CallGraphOptions = CallGraphOptions
+  { _graphFunctionNameFilter :: Maybe (NonEmpty Text)
+  }
+
+makeLenses ''CallGraphOptions
+
+parseCallGraph :: Parser CallGraphOptions
+parseCallGraph = do
+  _graphFunctionNameFilter <-
+    fmap msum . optional $
+      nonEmpty . Text.words
+        <$> option
+          str
+          ( long "function"
+              <> short 'f'
+              <> help "Only shows the specified function"
+          )
+  pure CallGraphOptions {..}
