+++ app/Commands/Extra.hs
@@ -1,47 +1,103 @@
 module Commands.Extra where
 
 import Juvix.Prelude hiding (Doc)
+import Control.Exception qualified as GHC
 import Options.Applicative hiding (hidden)
-import Options.Applicative.Builder.Internal
 import Options.Applicative.Types
+import System.Process
 
-hidden :: Bool -> Mod f a
-hidden sure = optionMod $ \p ->
-  if
-      | not sure -> p
-      | otherwise -> p {propVisibility = min Hidden (propVisibility p)}
-
-parserInputFile :: Parser FilePath
-parserInputFile =
+parseInputFile :: Parser FilePath
+parseInputFile =
   argument
     str
     ( metavar "JUVIX_FILE"
         <> help "Path to a .juvix file"
-        <> action "file"
+        <> completer juvixCompleter
     )
 
 parserInputFiles :: Parser [FilePath]
-parserInputFiles = many parserInputFile
+parserInputFiles = many parseInputFile
 
-addParser :: forall a c. Monoid a => Parser a -> Parser c -> Parser (a, c)
-addParser parser = \case
-  (NilP p) -> NilP ((\c -> (mempty, c)) <$> p)
-  (OptP (Option (CmdReader n cs g) ps)) ->
-    OptP (Option (CmdReader n cs (fmap optsInfo . g)) ps)
-    where
-      optsInfo :: ParserInfo b -> ParserInfo (a, b)
-      optsInfo pInfo = pInfo {infoParser = newParser}
-        where
-          newParser = do
-            opts <- parser
-            rest <- infoParser pInfo
-            pure (opts, rest)
-  (OptP o) -> OptP ((mempty,) <$> o)
-  (AltP p1 p2) -> AltP (addParser parser p1) (addParser parser p2)
-  (MultP p1 p2) ->
-    MultP
-      ((\(g2, f) (g1, x) -> (g1 <> g2, f x)) <$> addParser parser p1)
-      (addParser parser p2)
-  (BindP p k) -> BindP (addParser parser p) $ \(g1, x) ->
-    BindP (addParser parser $ k x) $ \(g2, x') ->
-      pure (g1 <> g2, x')
+juvixCompleter :: Completer
+juvixCompleter = Completer $ \word -> do
+  let cmd = unwords ["compgen", "-o", "plusdirs", "-f", "-X", "!*.juvix", "--", requote word]
+  result <- GHC.try @GHC.SomeException $ readProcess "bash" ["-c", cmd] ""
+  return . lines . fromRight [] $ result
+
+requote :: String -> String
+requote s =
+  let
+    -- Bash doesn't appear to allow "mixed" escaping
+    -- in bash completions. So we don't have to really
+    -- worry about people swapping between strong and
+    -- weak quotes.
+    unescaped =
+      case s of
+        -- It's already strongly quoted, so we
+        -- can use it mostly as is, but we must
+        -- ensure it's closed off at the end and
+        -- there's no single quotes in the
+        -- middle which might confuse bash.
+        ('\'': rs) -> unescapeN rs
+
+        -- We're weakly quoted.
+        ('"': rs)  -> unescapeD rs
+
+        -- We're not quoted at all.
+        -- We need to unescape some characters like
+        -- spaces and quotation marks.
+        elsewise   -> unescapeU elsewise
+  in
+    strong unescaped
+
+  where
+    strong :: String -> String
+    strong ss = '\'' : foldr go "'" ss
+      where
+        -- If there's a single quote inside the
+        -- command: exit from the strong quote and
+        -- emit it the quote escaped, then resume.
+        go '\'' t = "'\\''" ++ t
+        go h t    = h : t
+
+    -- Unescape a strongly quoted string
+    -- We have two recursive functions, as we
+    -- can enter and exit the strong escaping.
+    unescapeN = goX
+      where
+        goX ('\'' : xs) = goN xs
+        goX (x : xs) = x : goX xs
+        goX [] = []
+
+        goN ('\\' : '\'' : xs) = '\'' : goN xs
+        goN ('\'' : xs) = goX xs
+        goN (x : xs) = x : goN xs
+        goN [] = []
+
+    -- Unescape an unquoted string
+    unescapeU = goX
+      where
+        goX [] = []
+        goX ('\\' : x : xs) = x : goX xs
+        goX (x : xs) = x : goX xs
+
+    -- Unescape a weakly quoted string
+    unescapeD = goX
+      where
+        -- Reached an escape character
+        goX ('\\' : x : xs)
+          -- If it's true escapable, strip the
+          -- slashes, as we're going to strong
+          -- escape instead.
+          | x `elem` ("$`\"\\\n" :: String) = x : goX xs
+          | otherwise = '\\' : x : goX xs
+        -- We've ended quoted section, so we
+        -- don't recurse on goX, it's done.
+        goX ('"' : xs)
+          = xs
+        -- Not done, but not a special character
+        -- just continue the fold.
+        goX (x : xs)
+          = x : goX xs
+        goX []
+          = []
