+++ tests/CLI/Commands/dev/internal.test
@@ -1,0 +1,37 @@
+$ juvix dev internal
+>2 /Usage: juvix dev internal.*/
+>= 1
+
+$ juvix dev internal --help
+> /Usage: juvix dev internal COMMAND.*/
+>= 0
+
+$ juvix dev internal pretty
+> /Provide.*/
+>= 1
+
+$ juvix dev internal typecheck
+> /Provide.*/
+>= 1
+
+$ juvix dev internal typecheck tests/positive/Internal/Simple.juvix
+>
+Well done! It type checks
+>= 0
+
+$ juvix dev --only-errors internal typecheck tests/positive/Internal/Simple.juvix
+>
+>= 0
+
+$ juvix dev internal typecheck --only-errors tests/positive/Internal/Simple.juvix
+>
+>= 0
+
+$ juvix dev internal typecheck tests/positive/Internal/Simple.juvix --only-errors
+>
+>= 0
+
+$ juvix dev internal typecheck tests/negative/Internal/MultiWrongType.juvix --no-colors
+>2 /(.+)\/([^\/]+)\.juvix\:11\:7\-8\: error.*
+.*/
+>= 1
