+++ tests/CLI/Commands/dev/minic.test
@@ -1,0 +1,3 @@
+$ juvix dev minic
+> /Provide.*/
+>= 1
