+++ tests/CLI/Commands/dev/parse.test
@@ -1,0 +1,14 @@
+$ juvix dev parse
+> /Provide.*/
+>= 1
+
+$ juvix dev parse tests/positive/Axiom.juvix
+> /Module
+*/
+>2
+>= 0
+
+$ juvix dev parse --no-pretty-show tests/positive/Axiom.juvix
+> /Module \{.*/
+>2
+>= 0
