+++ tests/CLI/Commands/dev/root.test
@@ -1,3 +1,3 @@
-$ juvix internal root
+$ juvix dev root
 > /(.+)\/([^\/]+)/
 >= 0
