+++ tests/CLI/Commands/dev/scope.test
@@ -1,0 +1,3 @@
+$ juvix dev scope
+> /Provide.*/
+>= 1
