+++ tests/CLI/Commands/dev/termination.test
@@ -1,0 +1,23 @@
+$ juvix dev termination
+>2 /Usage: juvix dev termination COMMAND.*/
+>= 1
+
+$ juvix dev termination calls
+> /Provide.*/
+>2
+>=1
+
+$ juvix dev termination graph
+> /Provide.*/
+>2
+>= 1
+
+
+$ juvix dev termination calls --help
+> /Usage: juvix dev termination calls
+         \[\-f|\-\-function fun1 fun2 ...] \[\-d|\-\-show\-decreasing\-args ARG].*/
+>= 0
+
+$ juvix dev termination graph --help
+> /Usage: juvix dev termination graph \[\-f\|\-\-function ARG\].*/
+>= 0
