--- tests/CLI/Commands/internals/microjuvix.test
@@ -1,37 +1,0 @@
-$ juvix internal microjuvix
->2 /Usage: juvix internal microjuvix.*/
->= 1
-
-$ juvix internal microjuvix --help
-> /Usage: juvix internal microjuvix COMMAND.*/
->= 0
-
-$ juvix internal microjuvix pretty
-> /Provide.*/
->= 1
-
-$ juvix internal microjuvix typecheck
-> /Provide.*/
->= 1
-
-$ juvix internal microjuvix typecheck tests/positive/Internal/Simple.juvix
->
-Well done! It type checks
->= 0
-
-$ juvix internal --only-errors microjuvix typecheck tests/positive/Internal/Simple.juvix
->
->= 0
-
-$ juvix internal microjuvix typecheck --only-errors tests/positive/Internal/Simple.juvix
->
->= 0
-
-$ juvix internal microjuvix typecheck tests/positive/Internal/Simple.juvix --only-errors
->
->= 0
-
-$ juvix internal microjuvix typecheck tests/negative/Internal/MultiWrongType.juvix --no-colors
->2 /(.+)\/([^\/]+)\.juvix\:11\:7\-8\: error.*
-.*/
->= 1
