--- tests/CLI/Commands/internals/minic.test
@@ -1,3 +1,0 @@
-$ juvix internal minic
-> /Provide.*/
->= 1
