--- tests/CLI/Commands/internals/parse.test
@@ -1,14 +1,0 @@
-$ juvix internal parse
-> /Provide.*/
->= 1
-
-$ juvix internal parse tests/positive/Axiom.juvix
-> /Module
-*/
->2
->= 0
-
-$ juvix internal parse --no-pretty-show tests/positive/Axiom.juvix
-> /Module \{.*/
->2
->= 0
