--- tests/CLI/Commands/internals/scope.test
@@ -1,3 +1,0 @@
-$ juvix internal scope
-> /Provide.*/
->= 1
