--- tests/CLI/Commands/internals/termination.test
@@ -1,23 +1,0 @@
-$ juvix internal termination
->2 /Usage: juvix internal termination COMMAND.*/
->= 1
-
-$ juvix internal termination calls
-> /Provide.*/
->2
->=1
-
-$ juvix internal termination graph
-> /Provide.*/
->2
->= 1
-
-
-$ juvix internal termination calls --help
-> /Usage: juvix internal termination calls 
-         \[\-f|\-\-function fun1 fun2 ...] \[\-d|\-\-show\-decreasing\-args ARG].*/
->= 0
-
-$ juvix internal termination graph --help
-> /Usage: juvix internal termination graph \[\-f\|\-\-function ARG\].*/
->= 0
